backend/pyproject.toml
[project]
name = "deckify-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.18.4",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
    "taskipy>=1.14.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]
ignore = ["PLR0913", "PLR0917"]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]
omit = ["app/core/**", "app/main.py"]

[tool.taskipy.tasks]
dev = "fastapi dev app/main.py"
migrate = "alembic upgrade head"
lint = "mypy . && ruff check"
test-unit = "pytest -s -x --cov=app tests/unit -vv"
test-e2e = "pytest -s -x tests/e2e -vv"
test = "pytest -s -x --cov=app tests/unit -vv && pytest -s -x tests/e2e -vv"
format = "ruff format"
pre_format = "ruff check --fix"
post_test = "coverage html -d .htmlcov"
post_migrate = "format"
